backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
